requirements.txt
pyserial
boto3
python-dotenv